Move ControllerHelper functionality to BlogController
This both removes the need to even have the controller helper,
and makes more sense in terms of where it should live.

Now everything that inherits from BlogController can act like a
Blog Controller!
